fixes build progress state sync
Issue:
Sometimes sync plugin is not synchronizing the build
run state correctly. That's is resulting in intermittent
behavior that doesn't update waiting input state in OS build.
Such issue is occurring because `BuildSyncRunListener`  polls
Jenkins build run state periodically and update the status
in OS build on some conditions. Deriving such conditions
is kind of error-prone in some cases.
Also, polling at a fixed interval wasting some CUP cycles.

Fix:
Instead of polling based approach `BuildSyncRunListener`
listens the Jenkins run step events and updates status in
OS build resources.

Changes:
This commit updates existing listener to update build
status on workflow start, complete, finish, failed events.
Introduces a new listener that listens job steps and updates
build status post step execution event.
To serialize all build event updates (because events are coming
from two places)  `BuildStateUpdater` queues it and process it one by one.
Updates some test cases in favour of re-factoring.

Fixes
 - https://github.com/openshiftio/openshift.io/issues/3935
